refactor: extract delete thanks to @panarch
