+++ core/src/executor/execute.rs
@@ -4,21 +4,20 @@ use {
             alter_table, create_index, create_table, delete_function, drop_table, insert_function,
             CreateTableOptions,
         },
-        fetch::{fetch, fetch_columns},
+        delete::delete,
+        fetch::fetch,
         insert::insert,
         select::{select, select_with_labels},
         update::Update,
         validate::{validate_unique, ColumnValidation},
-        Referencing,
     },
     crate::{
         ast::{
-            AstLiteral, BinaryOperator, DataType, Dictionary, Expr, ForeignKey, Query,
-            ReferentialAction, SelectItem, SetExpr, Statement, TableAlias, TableFactor,
-            TableWithJoins, Variable,
+            AstLiteral, BinaryOperator, DataType, Dictionary, Expr, Query, SelectItem, SetExpr,
+            Statement, TableAlias, TableFactor, TableWithJoins, Variable,
         },
         data::{Key, Row, Schema, Value},
-        result::{Error, Result},
+        result::Result,
         store::{GStore, GStoreMut},
     },
     futures::stream::{StreamExt, TryStreamExt},
@@ -267,62 +266,7 @@ async fn execute_inner<T: GStore + GStoreMut>(
         Statement::Delete {
             table_name,
             selection,
-        } => {
-            let columns = fetch_columns(storage, table_name).await?.map(Rc::from);
-            let referencings = storage.fetch_referencings(table_name).await?;
-
-            let keys = fetch(storage, table_name, columns, selection.as_ref())
-                .await?
-                .into_stream()
-                .then(|item| async {
-                    let (key, row) = item?;
-
-                    for Referencing {
-                        table_name: referencing_table_name,
-                        foreign_key:
-                            ForeignKey {
-                                referencing_column_name,
-                                referenced_column_name,
-                                on_delete,
-                                ..
-                            },
-                    } in &referencings
-                    {
-                        let value = row.get_value(referenced_column_name).ok_or(
-                            ExecuteError::ValueNotFound(referenced_column_name.to_owned()),
-                        )?;
-
-                        let expr = &Expr::BinaryOp {
-                            left: Box::new(Expr::Identifier(referencing_column_name.clone())),
-                            op: BinaryOperator::Eq,
-                            right: Box::new(value.to_owned().try_into()?),
-                        };
-
-                        let columns = Some(vec![referencing_column_name.to_owned()]).map(Rc::from);
-                        let referencing_rows =
-                            fetch(storage, referencing_table_name, columns, Some(expr)).await?;
-
-                        let len = referencing_rows.count().await;
-                        if len > 0 && on_delete == &ReferentialAction::NoAction {
-                            return Err(ExecuteError::ReferencingColumnExists(format!(
-                                "{referencing_table_name}.{referencing_column_name}"
-                            ))
-                            .into());
-                        }
-                    }
-
-                    Ok::<_, Error>(key)
-                })
-                .try_collect::<Vec<_>>()
-                .await?;
-
-            let num_keys = keys.len();
-
-            storage
-                .delete_data(table_name, keys)
-                .await
-                .map(|_| Payload::Delete(num_keys))
-        }
+        } => delete(storage, table_name, selection).await,
 
         //- Selection
         Statement::Query(query) => {
